fix(sweep): type gap-heal cursor boundary args as bigint
CompleteGapHealCursor used NULLIF(sqlc.arg(observed_boundary_delivery_id), 0),
letting PostgreSQL infer the parameter as int4 and sqlc generate interface{}.
Delivery IDs above MaxInt32 then failed to encode (OID 23), leaving the active
gap cursor incomplete and retry jobs failing repeatedly (issue #46).

Cast the argument and sentinel to bigint so the generated param is int64.
Sentinel semantics unchanged: 0 still means no observed boundary. Regression
test completes a cursor with a delivery ID above MaxInt32 and asserts boundary
advance, pass-state clearing, and terminal completion.

Closes #46

+++ db/queries/sweep.sql
@@ -391,7 +391,7 @@ SET cursor = '',
     pass_high_watermark_at = NULL,
     boundary_delivery_id = COALESCE(
         NULLIF(pass_boundary_delivery_id, 0),
-        NULLIF(sqlc.arg(observed_boundary_delivery_id), 0),
+        NULLIF(sqlc.arg(observed_boundary_delivery_id)::bigint, 0::bigint),
         boundary_delivery_id
     ),
     pass_boundary_delivery_id = 0,
